Add login to generate api key, authentication component and get player route

+++ watchmelog/api/v1/views/players.py
@@ -1,22 +1,16 @@
-from apistar import types, validators
-from apistar.exceptions import BadRequest, NotFound
+import bcrypt
+from apistar.exceptions import BadRequest, NotFound, Forbidden
 from mongoengine import NotUniqueError
 
 from watchmelog.api.utils import hash_pass, mongo_to_dict
 from watchmelog.api.v1.models.players import Player, ApiKey
+from watchmelog.api.v1.types.players import RegisterPlayer, Login
 
 
-class RegisterPlayer(types.Type):
-    battletag: str = validators.String(description="Your Battle.net Tag")
-    password: str = validators.String(
-        min_length=12, description="Password used to authenticate with the API."
-    )
-    platform: str = validators.String(
-        enum=["PC", "XBOX", "PS4"], description="Platform you're playing on."
-    )
-    region: str = validators.String(
-        enum=["US", "EU", "ASIA"], description="Region you're playing on."
-    )
+def get_player(player_slug: str, auth_player: Player) -> dict:
+    if auth_player.slug != player_slug:
+        raise Forbidden("You cannot access that player.")
+    return mongo_to_dict(auth_player)
 
 
 def register_player(player: RegisterPlayer) -> dict:
@@ -32,19 +26,32 @@ def register_player(player: RegisterPlayer) -> dict:
             detail=f"A player with the battletag {player.battletag} already exists."
         )
 
-    return mongo_to_dict(new_player, black_list=["password"])
+    return mongo_to_dict(new_player)
 
 
-def generate_api_key(player_slug: str) -> dict:
+def generate_api_key(player_slug: str, login_payload: Login) -> dict:
+    """
+    Generate a new API Key for the player.
+    If an old key already exist, it will be deleted.
+    """
     try:
         player = Player.objects(slug=player_slug)[0]
     except IndexError:
         raise NotFound(f"Player {player_slug} not found.")
 
+    if (
+        not bcrypt.hashpw(
+            login_payload.password.encode("utf-8"), player.password.encode("utf-8")
+        ).decode("utf-8")
+        == player.password
+    ):
+        raise Forbidden("Wrong password")
+
     old_api_key = ApiKey.objects(player=player)
     if old_api_key:
         for key in old_api_key:
             key.delete()
+
     new_api_key = ApiKey(player=player)
     new_api_key.save()
     return mongo_to_dict(new_api_key)
